Refactor transpose handling and enhance MIDI loader UI
This commit updates the transpose handling in the engine to support signed values, allowing for a range of -24 to +24 semitones. The `get_pattern_transpose` method now returns an `int8_t`, and related calculations throughout the MIDI processing functions have been adjusted to accommodate this change. Additionally, the web editor has been enhanced with a new MIDI loader interface, allowing users to load MIDI files and manipulate note patterns more effectively. The transpose input field in the editor has also been updated to reflect the new signed range.

Files modified: src/engine.h, src/main.cpp, src/midi.cpp, src/midi_api.h, tools/web-pattern-edit/index.html, and platformio.ini.

+++ src/engine.h
@@ -1197,8 +1197,9 @@ struct Engine {
   uint8_t get_midi_note() const {
     return uint8_t(36 + get_sequence().get_pitch_dir(last_step_dir_));
   }
-  uint8_t get_pattern_transpose() const {
-    return get_sequence().transpose;
+  // Signed pattern transpose (-24..+24, 0 = none). Stored as int8 in the byte.
+  int8_t get_pattern_transpose() const {
+    return int8_t(get_sequence().transpose);
   }
   bool get_slide() const {
     return get_sequence().get_slide();
